Mechanika gry - w toku

+++ EduBot/EduApi/EduApi/user_game.cs
@@ -16,7 +16,7 @@ namespace EduApi
     {
         public int user_id { get; set; }
         public int life { get; set; }
-        public int shield { get; set; }
+        public decimal shield { get; set; }
         public int rank { get; set; }
     
         public virtual user user { get; set; }
